+++ Assets/Scripts/GameManager.cs
@@ -4,12 +4,6 @@ using UnityEngine;
 
 public class GameManager : MonoBehaviour
 {
-    // Concept:
-    // Phase 1: Cities have different requests, and conflict is kept to a minimum
-    // Phase 2: Cities have the same request
-    public enum Phase { Phase1, Phase2 };
-
-
     [Serializable]
     public class PlayerData
     {
@@ -19,7 +13,7 @@ public class GameManager : MonoBehaviour
         public int      score;
     }
 
-    [SerializeField] private int                _numPlayers = 1;
+    [SerializeField] private int                _numPlayers = 2;
     [SerializeField] private List<PlayerData>   _playerData;
 
     private float playTime = 0.0f;
@@ -76,13 +70,6 @@ public class GameManager : MonoBehaviour
     private void Update()
     {
         playTime += Time.deltaTime;
-    }
-
-    public static Phase GetPhase()
-    {
-        if (Instance.playTime < 30.0f) return Phase.Phase1;
-
-        return Phase.Phase2;
     }
 
     public int numPlayers