+++ RELEASE_NOTES.md
@@ -1,5 +1,12 @@
 # RELEASE NOTES
 
-## Version 1.0
+## Version 1.0 (26/01/2025)
 
-- Initial release
+- Initial release (Global Game Jam'25 version)
+
+## Version 1.1 (27/01/2025)
+
+- Improved submarine
+- Added phase display
+- Phase 1 and 2 now don't allow for torpedos
+- New phase 3, same as 2 but with weapons